Feat add confirm message for product delete and add code so that only super users can update the FAQ

+++ products/views.py
@@ -225,9 +225,10 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
 
-# @login_require
-# def delete_image
+    return render(request, 'products/product_confirm_delete.html', {'product': product})
